Update download_data.py
Changed info.party in makegraph() to store data in a list rather than a set. This allows the user to calculate the mode of the party to see what the most common party affiliation is. 

Also, created a allownegweights flag in makegraph() that defaults to False, so one can continue using the prior version without change. If one sets allownegweights to True, the weights in the graph will be the cosine similarity between two vertices, which can be negative.